+++ src/main/java/me/ajlane/geo/repast/succession/AliasedEnvrConsequent.java
@@ -1,8 +1,0 @@
-package me.ajlane.geo.repast.succession;
-
-public class AliasedEnvrConsequent extends EnvrConsequent<String, Integer> {
-
-  public AliasedEnvrConsequent(String targetState, Integer transitionTime) {
-    super(targetState, transitionTime);
-  }
-}