+++ src/main/java/me/ajlane/geo/repast/succession/CodedEnvrConsequent.java
@@ -1,8 +1,0 @@
-package me.ajlane.geo.repast.succession;
-
-public class CodedEnvrConsequent extends EnvrConsequent<Integer, Integer> {
-
-  public CodedEnvrConsequent(Integer targetState, Integer transitionTime) {
-    super(targetState, transitionTime);
-  }
-}