+++ src/main/java/me/ajlane/geo/repast/succession/LcsTransitionMapFactory.java
@@ -1,6 +1,0 @@
-package me.ajlane.geo.repast.succession;
-
-public interface LcsTransitionMapFactory {
-  CodedLcsTransitionMap getCodedLcsTransitionMap();
-  AliasedLcsTransitionMap getAliasedLcsTransitionMap();
-}