+++ src/main/java/me/ajlane/geo/repast/succession/pathway/convert/EnvrConsequentConverter.java
@@ -1,0 +1,40 @@
+/**
+ * 
+ */
+package me.ajlane.geo.repast.succession.pathway.convert;
+
+import me.ajlane.geo.repast.succession.pathway.aliased.AliasedEnvrConsequent;
+import me.ajlane.geo.repast.succession.pathway.coded.CodedEnvrConsequent;
+
+/**
+ * Convert a CodedEnvrConsequent to an AliasedEnvrConsequent or vice versa
+ * 
+ * @author Andrew Lane
+ *
+ */
+public class EnvrConsequentConverter {
+  EnvrStateAliasTranslator translator;
+
+  EnvrConsequentConverter(EnvrStateAliasTranslator translator) {
+    this.translator = translator;
+  }
+
+  CodedEnvrConsequent convert(AliasedEnvrConsequent cons) {
+    EnvrStateAliasTranslator tr = this.translator;
+    CodedEnvrConsequent convCons =
+        new CodedEnvrConsequent(tr.numericalValueFromAlias("landCoverState", cons.getTargetState()),
+            cons.getTransitionTime());
+
+    return convCons;
+  }
+
+  AliasedEnvrConsequent convert(CodedEnvrConsequent cons) {
+    EnvrStateAliasTranslator tr = this.translator;
+    AliasedEnvrConsequent convCons = new AliasedEnvrConsequent(
+        tr.aliasFromNumericalValue("landCoverState", cons.getTargetState()),
+        cons.getTransitionTime());
+
+    return convCons;
+  }
+
+}