+++ src/main/java/me/ajlane/geo/repast/succession/pathway/convert/package-info.java
@@ -1,0 +1,16 @@
+/**
+ * Convert between different representations of succession pathways
+ *
+ * <p>
+ * Succession pathways can be expressed in terms of either numerical codes or natural languages
+ * aliases. This is to enable modellers to represent succession processes in terms they understand
+ * (natural language) and automatically covert this information into numerical data that can be
+ * easily and efficiently manipulated in a simulation model. Classes in this package are used to
+ * convert between the two representations.
+ * </p>
+ *
+ * @see me.ajlane.geo.repast.succession.pathway.aliased aliased
+ * @see me.ajlane.geo.repast.succession.pathway.coded coded
+ * @author Andrew Lane
+ */
+package me.ajlane.geo.repast.succession.pathway.convert;