+++ src/main/java/me/ajlane/geo/repast/succession/pathway/io/package-info.java
@@ -1,0 +1,10 @@
+/**
+ * Load specifications of ecological succession pathways from persistent storage
+ *
+ * <p>
+ * The specification of ecological transitions allowed in the model is treated as a separate step
+ * independent of the specification of the AgroSuccess simulation model. Here we include code used
+ * to read these ecological succession models from disk to incorporate them in simulations.
+ * </p>
+ */
+package me.ajlane.geo.repast.succession.pathway.io;