+++ src/test/java/me/ajlane/geo/repast/succession/EnvrConsequentTest.java
@@ -3,6 +3,7 @@ package me.ajlane.geo.repast.succession;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 import org.junit.Test;
+import me.ajlane.geo.repast.succession.pathway.EnvrConsequent;
 
 public class EnvrConsequentTest {